Update README.md
Cloning with ssh is not possible without adding an SSH key. So for the general public it is better when you add cloning with HTTPS. Furthermore you need to either add a "dot" at the end of the clone command or then add "cd graphsense-setup" as a second command to your example.